[K/N] Implement protocol type checks without using protocol getters
See KT-79741 for the context.
Building on the previous commits, this commit introduces an alternative
implementation for type-checking Kotlin objects against Objective-C
protocols (`obj is FooProtocol`).

To select between the old and the new implementation, the compiler
checks the `cCallMode` binary option, because cinterop does the same --
piggybacks the `-Xccall-mode` machinery, following its general meaning.

The old implementation is pretty straightforward: given a Kotlin object,
it gets its Objective-C peer (this operation is defined for any Kotlin
object). Then it gets the protocol getter name from the class
annotation, calls the getter to get the protocol pointer, and invokes a
runtime function (`Kotlin_Interop_DoesObjectConformToProtocol`) that
checks whether the given Objective-C object conforms to the given
protocol.

Meta-classes are also handled there: Objective-C runtime considers
the class object conforming to the protocol as well, so the Kotlin
runtime type check implementation distinguishes regular objects from
class objects to indicate that regular objects conform only to regular
protocol interfaces, and class objects conform only to "meta"
interfaces.

The new implementation goes similarly: it introduces the new runtime
function, `Kotlin_Interop_DoesObjectConformToProtocolByName`, that takes
the protocol name instead of a pointer. Compiler extracts this name from
the `binaryName` parameter of the class annotation.
So, the runtime function then requests Objective-C runtime to find the
protocol by its name using `objc_getProtocol`, and then does the same
check as the old implementation.
To mitigate performance costs of the name lookup, the runtime function
caches its result using the global provided by the compiler.

So far it is pretty straightforward. Here go the tricks.

There is a special case -- Kotlin subclasses of Objective-C classes
(known as "Kotlin Objective-C classes" aka `IsKotlinObjCClass`).
Instances of such classes mostly behave like instances of regular
Objective-C classes, but their Kotlin peers also contain Kotlin fields.
The Objective-C classes are created at runtime. To not bother with
Objective-C protocols ABI, the compiler puts only protocol names to
`KotlinObjCClassInfo`, and the runtime looks up protocols by name and
makes the Objective-C classes it creates adopt them.
But there is a catch: if it doesn't find a protocol by name, it doesn't
add it. And that can legitimately happen.

Objective-C protocol runtime information is a tricky thing: Clang
generates them on demand as weak globals. But the Kotlin compiler
doesn't generate them, to not deal with the complexity (ironic).
So, a protocol implemented in Kotlin can lack runtime information,
leading to `objc_getProtocol == nullptr` for it.

Worth mentioning that before the `-Xccall-mode direct`, this could never
happen with protocols imported to Kotlin: cinterop always generated
protocol getters which basically returned `@protocol` references.
Even if a protocol getter was DCEd, the protocol runtime data was not.

So, the whole problematic case for the new type check implementation is
the following:
1. There is an Objective-C protocol that doesn't have any
   implementations or `@protocol()` references (but can have callsites).
   It is processed by cinterop with `-Xccall-mode direct`.
2. A Kotlin class implements this protocol. Other Kotlin code checks
   whether an instance of this class conforms to that protocol.
3. Since the protocol doesn't have the runtime information, the Kotlin
   runtime doesn't find it when creating the Objective-C class and
   doesn't add the protocol to the class.
4. When performing the type check, it also can't find the protocol and
   can't perform the type check.

Here is how this commit handles this case.
When the runtime function can't find the protocol, it checks whether
this object is a Kotlin Objective-C class instance by checking the
`TypeInfo` flag, `TF_KOTLIN_OBJC_CLASS`.
If it is, the runtime then looks up for `KotlinObjCClassInfo` for that
object and uses the list of protocol names from it. It then simply
checks whether the list has the protocol name in question.

^KT-79741 Fixed