[FIR] Refactor frontend interpreter on shared evaluation map
The frontend and backend interpreters now use the same map to evaluate
builtin functions. This removes the almost semantically equivalent
OperationsMap and it's generator.

For K1 some special handling was added as it uses the evaluator also for
typechecking that would lead to K1 no longer throwing type checking
errors for some invalid expressions.

^KT-51065