[Parser] Fix infinite loop in destructuring full form parsing
The logic for detecting the following situation

val
a

works by rolling back to before the name identifier and relied on
a val or var token having been consumed before to prevent and endless
loop.

In the case of an incomplete full form destructuring declaration `(val`,
there is no val or var before the parentheses and so nothing was
consumed, the parser rolled back to before the declaration and an
endless loop began.

#KT-79951 Fixed