Fix mixing relative friend paths and absolute classpaths with symlinks
The fix for KT-60886 (a5b530d5) converted calls to toAbsolutePath()
to toRealPath() to fix the testDirectoryWithRelativePath test, which
was failing due to comparing a normalized path to an unnormalized
path (with ../../.. in it).

Calling toRealPath() instead of toAbsolutePath() to get a normalized
path also resolves any symlinks in the path to their targets.  Since
toRealPath() is not called on the absolute path it is being compared
to it results in different absolute paths and the match check always
fails.

Use toAbsolutePath().normalize() to produce a normalize absolute path
instead.

Fixes https://youtrack.jetbrains.com/issue/KT-80039