K2: take arrays into account while checking for MISSING_DEPENDENCY_CLASS
During the fix of KT-79866, it was decided not to check types of
function parameters on missing dependencies in case a default argument
is in use. The reason was a number of hits like this in intellij project
and an opinion that types of such parameters are effectively unused.
However, a set of factors forces us to check them anyway if a parameter
type is an array: generic arguments are approximated in FIR2IR,
and approximation of inaccessible types leads to IrErrorType creation
for these arguments even if approximation is effectively not needed;
then type arguments are normally erased in JVM codegen, but arrays
are an exception. So to avoid potential problems with NonExistentClass
in bytecode, we have to forbid this specific arrays case back.

#KT-79866 Fixed

(cherry picked from commit 5c41dcaadfa42d2e6bf0f19d38f878be23bdc950)