[K/N] Mention that `ExperimentalForeignApi` is Beta
Currently, the KDoc for the `kotlinx.cinterop.ExperimentalForeignApi`
annotation mentions that the marked API is "experimental".

This opt-in is required for certain things related to C and Objective-C
interoperability:
* Some API in `kotlinx.cinterop.*`, required when working with native
  libraries.
* All declarations in native libraries, except for "platform libraries".

But the website claims that C and Objective-C interoperability is Beta:
https://kotlinlang.org/docs/components-stability.html#kotlin-native
This doesn't match the "experimental" in the opt-in KDoc.

This commit fixes this confusion by making the KDoc for the opt-in also
mention the Beta status.

^KT-77441 Fixed