[JS] Copy JS export model classes to the new jsexport package
We are going to have two export models: one for generating
.d.ts files and one for generating the actual JS statements for
exported declarations.
The latter will contain the corresponding IR nodes, and
the former will be independent of IR.
Eventually, the former will be constructed from the
frontend metadata via Analysis API.

This commit just copies the classes with no changes.
In the next commits we will simplify the new export model to only
contain the information necessary for generating the JS statements.

KT-79628