[IR] Ignore invalid IR from scripting plugin
Scripting is the only compiler mode/plugin that breaks that many "rules"
of IR. The invalid IR comes from both FIR2IR, if there are script sources
(because of the IrReplSnippet element), and from the compiler plugin,
which then lowers IrReplSnippet into classes.
To not fail in the scripting tests, but also to not disable IR
validation completely, we disable only the affected checkers, only under
scripting related IR elements.

KT-79807
KT-80071
KT-80062