[Inliner] Don't capture variables outside inline lambda in locals
When doing small application of LocalDeclarationsLowering inside
inline lambdas, we don't plan to popup classes outside of the current
function, we only move them outside of call.

This makes capturing values defined in the current scope unneeded.
While doing so is correct, it can be suboptimal, and for example lead to
capturing crossinline lambda to field, which prevents it's further
inlining by IR inliner.

To avoid this, we we now only capture variables, which are not defined
in the scope, where we want to put class.

^KT-69497 Fixed