[FIR2IR] Improve reporting in IR validation with more than one plugin
Consider this scenario:
1. Plugin A runs and generates some invalid IR.
2. We validate IR and report some violations were caused by plugin A.
3. Plugin B runs and does nothing.
4. We validate the same IR once again and report the same violations as
caused by plugin B.

This would be incorrect, as the plugin B is innocent here. If someone
reads the build log bottom-up, they may get the wrong blame.

The simplest and the chosen solution is to just suggest in the message
that the violation may be caused by some earlier plugin.

Alternative 1: Mark which IR element was created by which plugin. But
we also have to account for modifying existing elements, that is very
hard technically.
Alternative 2: "Deduplicate" the reported violations. That is fragile,
e.g. what if plugin B calls deepCopy() on the invalid IR? deepCopy
itself is rather innocent operation.

Even more problematic is if FIR2IR produced an IR that causes violations
to begin with. That _shouldn't_ happen, this is also caught as an error
in our tests, but if it does, we ought not to blame compiler plugins.

KT-77819
KT-80070