[AA REx] Require a parent Disposable when creating resolve extensions.
Per the documented API for Resolve Extensions, the `KaResolveExtension`
instance is `Disposable`, and will be cleaned up by the creator when it
is no longer necessary. This was handled by `LLFirSession` for the
extensions created for analysis, but the extensions created by
`KaResolveExtensionToContentScopeRefinerBridge` did not dispose their
created extensions, causing resource leaks in REx clients.

To fix this, and to enforce proper handling of the `Disposable`
lifecycle for `KaResolveExtension` users, require passing a `Disposable`
factory into `provideExtensionsFor`, which will be used as a parent
disposable for any created extensions. For efficiency, the factory will
not be called if no extensions are created.