[Gradle] Request artifactType in Project shared data API explicitly
This should improve accuracy of Gradle Artifact Transformation algorithm
As it based of BFS on graph of transformations and it starts backwards
i.e. from "requested" attributes to find an appropriate chain of
transformations that would lead to attributes of consuming artifact.

and since a lot of artifact transformations are defined like
Transform from artifactType=X to artifactType=Y. other attributes are
not mentioned at all. This leads into compatibility with requested
attributes that have no artifactTypes. This is not desirable and
may lead to memory explosion when building up BFS front.

^KT-76265 In Progress