[ABI Validation] Embedded intellij deps asm-all into abi-tools
Artifact org.jetbrains.intellij.deps:asm-all is published only in Intellij Deps repository.
Since the checkLegacyAbi and updateLegacyAbi updateLegacyAbi require this dependency, this situation is most inconvenient for the user.

It was decided to embed this dependency directly into `abi-tools` so as not to force users to add an unfamiliar repository


Merge-request: KT-MR-23039
Merged-by: Sergei Shanshin <Sergey.Shanshin@jetbrains.com>