[FIR] Fix mapping collection for KMP separate compilation
We need to be able to actualize to the same platform class from any of
its common variants.
In the test `fakeOverrideExpectSuperTypeFromLib`, this was failing
because we only computed the mapping from the intermediate source set
while missing the one from the common one.

The reason was that when a FirCommonDeclarationsMappingSymbolProvider
from a common session (e.g. commonMain) only got one symbol, it didn't
save this information. In many cases, the mapping would later be
created in `referenceAllCommonDependencies`. However, it only worked
when the intermediate source set saw the same variant of the class
as the common one.

After this commit, the symbol provider remembers any classes so that
the mappings can be computed in
IrCommonToPlatformDependencyActualizerMapContributor.

#KT-80131 Fixed
#KT-77385 Fixed
#KT-80002 Fixed