Enable use of deploy version for Kotlin Native in TeamCity
With this, versions.kotlin-native is set to the version of artifacts
in current build by default.
The tests that used the fallback version of Kotlin Native are switched
to work with Kotlin Native from local repo set up by TeamCity.
There are tests that rely on Kotlin Native being unpacked
in the project configuration stage. This is broken when using Maven repo
so a workaround is applied.

This is reapplication of commit b08685cd32b27f10ec449e9dce3c4c2bc0fcbbc5
with additional minor changes.

KTI-2342
KTI-2238