[Inline] Fix computing signature of prepare inline function copy
When there is a reference from one prepared inline function to another,
NonLinkingIrInlineFunctionDeserializer may link the symbol to the
prepared copy, rather than the original function, because it uses a
separate symbol table that does not know about the original functions.
It was rare or impossible before, but now it does occasionally occur,
when all prepared inline functions are stored inside a single file.

The problem is when trying to serialize such symbol later, because the
prepared function copy itself does not always store enough data to
compute the signature for it. So we have to look for the original one.

KT-75794