[Gradle] Emit diagnostic on unsupported AV/LV with `kotlin-dsl`
This change adds an additional diagnostic from the tooling side if we detect unsupported API or language version used in a project with `kotlin-dsl` plugin. `kotlin-dsl` configures those to specific values forcibly and with a reason. Using non-embedded version of Kotlin compiler in a build logic build is often a mistake leading to different kinds of issues. This particular case is among them and is considered to be seen relatively often, so we cover it by a diagnostic giving special recommendation with regards to `kotlin-dsl` in addition to warnings/errors from the compiler.
^KT-79851 Verification Pending