[stdlib] Add @UsedFromCompilerGeneratedCode annotation
Sometimes it's hard to tell whether it's safe to refactor an internal
declaration inside the standard library (e.g., rename or remove it)
without breaking the compiler.
Sometimes we mark such declarations with
`@Suppress("unused") // usages are generated by the compiler`,
but other times we also have to use such a declaration inside
the standard library, which makes `@Suppress("unused")` kind of
confusing.

We solve this by introducing the new
`kotlin.internal.UsedFromCompilerGeneratedCode` annotation, with which
we mark the internal stdlib declarations that are used from
the compiler.

Later we plan to introduce a logic for enforcing the presence of this
annotation if the declaration is being loaded in the compiler
via `org.jetbrains.kotlin.ir.SymbolFinder`.

Also, we will configure the IDE to treat annotated declarations as
entry points, so it doesn't highlight them as unused.

KT-78572