KT-71822 Align Iterable.intersect/union/subtract behavior with a contract
For certain inputs, `Iterable<T>.intersect` (as well as `union` and `subtract`)
were not obeying the contract. Primarily, it was a case when collections
with a `contains` that use anything but structural equality (for instance, identity sets).

This change aligns the behavior with a contract, but also clarifies the contract for edge
cases.

^KT-71822 Fixed

Merge-request: KT-MR-18207
Merged-by: Filipp Zhinkin <filipp.zhinkin@jetbrains.com>