Adapt PluginStructureProvider to 2025.2 ahead
In 252, PathResolver#resolvePath dropped readInto parameter, mentioning
that it's not used. Indeed, even with 241 for now, passing a default
RawPluginDescriptor instance is no-op (or same as passing null).

In addition, the return value of resolvePath call is not used at the
moment (violating check-return-value). In 252, that read XML content
isn't quite used if we return the default RawPluginDescriptor instance
again. Instead, the original intention is supposed to be elvis, i.e.,
try resolvePath or use a default instance.

^KT-79870