[FIR2IR] Use memberRequiredPhase = STATUS in scopes
This fixes a kotlin.git specific bug in evaluator.
Constant evaluation accesses the declarations of lazy class KFunction
when evaluating KCallable.name which builds a scope and runs into this
problem inside fake override generation. This can only really happen
when KCallable is a source class. In any other project, KCallable will
come from a binary dependency.

Apart from that, it's seem to be basically impossible to construct a
scenario where we build a scope for some class for the first time in
FIR2IR. If it's asource class, we certainly would have built a scope
with `requiredPhase = STATUS` during the checkers phase.

#KT-79887 Fixed