[Commonizer] Don't account for `isValue` when setting `hasAnnotations`
...of the metadata... built from the commonized Cir. With `FirElementSerializer`,
the flag is only set if there are actually some annotations, so we must
leave the metadata in the same state as they had when we accepted it as
the input.

Without this change, some test fail complaining that the commonization
result contains unexpected `true` values for `hasAnnotation` when compared
to the golden files.
See: `ClassifierCommonizationFromSourcesTest#testClassKindAndModifiers`.