[LL] BodyStateKeepers: properly guard property initializers and delegates
The previous solution has a bug – it stored the `null` value as a value
to preserve in the case of exceptions. This led to a situation where
an already resolved initializer was dropped due to an exception from
the getter.
The proper solution is to not guard already resolved value at all

^KT-75858 Fixed