[K/N] Support generating `@CCall.Direct` in cinterop
See KT-79752 for the context.

This commit introduces a new flag to cinterop: `-Xccall-mode`
The default value is `indirect` (so the current behavior remains
unchanged so far).

When passing `direct`/`both`, cinterop generates `@CCall.Direct`
annotations instead of/in addition to `@CCall` correspondingly.

C global accessors are also imported to Kotlin via `@CCall`, but they
can't support `@CCall.Direct`, so they are instead marked as not yet
supported when the `direct` mode is specified.

^KT-79752