[K/N] Don't pack bitcode to cinterop klibs with `-Xccall-mode direct`
See KT-79752 for the context.

The previous commits introduce the generation of `@CCall.Direct` in
cinterop when `-Xccall-mode` is specified.

This commit disables including any bitcode to a cinterop klib when it
is generated with `-Xccall-mode direct`.

Therefore, a cinterop klib generated with `-Xccall-mode direct` doesn't
require the K/N compiler to use a compatible LLVM version, enabling
compatibility with older compiler versions (if they support
`@CCall.Direct`).

Additionally, this commit prohibits using `entryPoint=` in `.def` files
and `-Xcompile-source` cinterop flag if `-Xccall-mode` is not `indirect`
(i.e. not default).
See the explanation for this decision in the source code.

^KT-79752