[K/N] Support generating calls to `@CCall.Direct` in the compiler
See KT-79753 for the context.

This commit makes the compiler able to generate calls to functions
marked with `@CCall.Direct`.

It introduces a new binary option

    cCallMode=indirect|indirectOrDirect|directOrIndirect|direct

that specifies whether the compiler should support `@CCall` and
`@CCall.Direct` and which one it should prefer if a function is marked
with both.

The default value is `indirectOrDirect`, i.e. use `@CCall` unless only
`@CCall.Direct` is available (which can happen if a cinterop klib is
generated with `-Xccall-mode direct`).

^KT-79753