[K/N][tests] Don't use -Xcompile-source for cinterop modules
The K/N test data files can define modules with `// MODULE` directive.
Apart from a regular Kotlin module, such a module can be a cinterop
module (distinguished by a `.def` file among its files).

cinterop modules in tests can also define source files in C, C++ or
Objective-C(++). When running such a test, the test infrastructure
delegates compiling such source files to the cinterop tool itself by
passing the -Xcompile-source flag to it.

The problem with that approach is that the source files are compiled to
bitcode, which is then included in the resulting cinterop klib.
At the same time, KT-79751 introduces a new mode where cinterop klibs
don't have bitcode inside.

Therefore, the approach with -Xcompile-source can't be used to test this
new mode.

To address this, this commit reimplements the support for source files
in cinterop modules in tests: instead of using -Xcompile-source, the
source files are compiled to static libraries and then packed to
cinterop klibs using the -staticLibrary flag.

As a side effect, having native code compiled to static libraries
instead of bitcode makes the corresponding cinterop tests closer to the
real-world scenarios.

^KT-79753