[Analysis API Standalone] KotlinStandaloneIndexBuilder: cache files between tests
It is possible to reuse a pair of a decompiler and a virtual file to
"share" `KtFile` between different projects. This helps to avoid
costly decompiler checks since even with cached content for virtual
files, they still have an overhead. Also, it allows avoiding redundant
libraries traversal over and over again since the cache stores values
for all files inside a binary root

^KT-80559 Fixed