[K/N][tests] Remove some dead code to fix two tests after 377465a
377465a changes the way native source files are handled in cinterop
modules in tests: before the change, they were handled with the
cinterop flag `-Xcompile-source` which makes them compile to bitcode.
After the change, they are compiled to regular static libraries.

This broke a pattern used in certain tests: a runtime function called
from such a source file. Before the change, everything was compiled and
linked as bitcode. After the change, the runtime function is
internalized and DCEd first, and the usage is linked only after the
runtime is compiled to machine code, when the function is already gone.

377465a fixed that in most cases. But there are two more tests with the
same problem which didn't fail in the Aggregate CI configuration
because they fail only on non-Apple platforms: the problematic code is
dead, and it gets removed with `-dead_strip` on Apple platforms.

This commit fixes the tests by removing the dead code.