IR Inliner: Soft-extraction of local delegated properties from inlinable lambdas
With this commit we are adding soft-extraction of local delegated
properties from inlinable lambdas: The `delegate` variable stays
inside the lambda while the rest of `IrLocalDelegatedProperty` is
moved one level above -- to keep it outside of the lambda.

This is necessary since there may be references to LDPs from local
classes and local functions, which are also extracted from inlinable
lambdas.

And we cannot get rid of LDP completely on the first phase, because
there are still references to LDP symbol (see the details in KT-77103).

^KT-78856
^KT-77103 Fixed