[fir2ir] Treat code fragments returning 'Nothing' as non-void
By default, the JVM backend maps the 'Nothing' return type to 'Unit'
(see 'JavaToKotlinClassMap.mapKotlinToJava'). For code fragments, such
behavior is rather unexpected, especially if the 'Nothing' type appeared
as a result of a smart cast.

The change replaces the return type of 'Nothing'-returning fragments
to 'Any?' so it's mapped to debugger-friendly 'Object' instead.

^KT-70860 Fixed