[FIR] Report missing diagnostics on delegated properties with accessors
Delegated properties are allowed to have accessors as long as they
don't have a body, e.g., to put annotations on them.

In this situation, we would previously set the source kind of the FIR
accessor to DelegatedPropertyAccessor. This would make it impossible to
distinguish between a delegated property with accessor but without body
and a delegated property without accessor. The case with accessor with
body was always possible to recognize.

After this commit, the accessor gets a real source kind if it was
defined in source. To recognize that it's still a generated accessor
for a delegated property, we need to check the source kind of the first
body statement. The helper `hasGeneratedDelegateBody` is introduced for
this purpose.

#KT-80600 Fixed