[ABI Validation] Added dependency on checkKotlinAbi from `check` task
Now, even if ABI validation is enabled in the Gradle project, it is not called automatically when running the check task.

This is counterintuitive, at least because the task is also called checkKotlinAbi.

The check task description also says: Runs all checks. Since checkKotlinAbi is a checking task, we have to run it too.

#KT-80614 Fixed