[Gradle] Test for the correct message in appleGradlePlugin IT
Previously resolving iosAppIosX64ReleaseImplementation0 configuration
for iosLib dependency were resulting in *sources* variant.
This is 100% incorrect behavior, with recent changes as part of this
KT-76265. artifactType attribute value was changed. And that
caused now correct behavior of "failing" the resolution.
Because iosLib doesn't expose Release framework. And there is no
compatibility variant for apple binary types debug / release.

So this is expected to fail!

^KT-80720 Verification Pending