[FIR] Fix argument mapping of invoke calls with context arguments
Argument lists entries must be unique because we build a
Map<FirExpression, FirValueParameter> for the argument mapping.
But in `invoke` calls, we convert context arguments to regular
arguments, and because expressions are created once per implicit value,
we can get duplicates.

To fix this, create expression copies when we convert context arguments
to regular arguments.

#KT-77541 Fixed