[K/N] Don't fail on non-existent transitive klib dependency in cinterop
cinterop accepts klib dependencies via CLI flags `-library`/`-l`.
It uses a standard klib resolver which fails whenever a klib manifest
mentions a non-existent dependency in `depends =`.

This behavior is not relevant anymore, since Gradle is supposed to take
care of transitive dependencies itself.
And such failures are also undesirable in practice, see KT-80593.

The compiler used to be affected by the same problem, but it was fixed
in 1f3b1f4.
This commit, in turn, fixes it in cinterop by flipping the same flag.

^KT-80593 Fixed