[Build] Update LV and AV for `kotlin-native/performance/buildSrc` to 2.3
`kotlin-dsl` Gradle plugin, used in that project, sets the language and
api versions to 1.8. But we compile it with the Kotlin bootstrap
compiler, which doesn't support LV/AV=1.8 anymore.

Previously, 774614f worked this around by setting LV and AV to 1.9.
But this version is also going to be prohibited soon.
So, this commit, in turn, bumps the versions to 2.3.

^KT-80147 Fixed