Reflection: use descriptor-less KProperty for local delegated properties
Local delegated properties are the simplest kind of declarations which
we can reimplement using kotlin-metadata-jvm and check that basic things
are working: they always have default accessors, no type parameters from
outer declarations, call/callBy are not supposed to work, etc.

In this change, the general hierarchy of new Km-based declarations is
introduced (KotlinKCallable, KotlinKProperty, etc), however many methods
are not implemented properly yet because it's not needed for local
delegated properties. These methods call an assertion
`checkLocalDelegatedPropertyOrAccessor` which will be removed in future
commits where the new objects are supported for other kinds of
callables.

 #KT-80710