[FIR] Add REPL declaration reference FIR node
REPL snippets will soon behave similar to both class initialization and
a function body. A REPL snippet body is wrapped in a function which
itself is placed within an object. Non-nested declarations (properties,
functions, classes, etc.) are extracted from the function body and
placed within the object. To make sure resolution is performed in the
correct order, this extract leaves behind a reference to the new
class-level declaration in the REPL snippet function body.

^KT-82575