[JS] Introduce an entry point for Analysis API-based TypeScript export
This is heavily inspired by Swift Export.

Currently, generating the export model itself is not implemented;
there are only some stubs.
It will be implemented in later commits.

The entry point is the `runTypeScriptExport` function, similar to
`runSwiftExport`. It is supposed to be called from KGP and from compiler
tests.

Test runners for the new TS export are added in the next commit.
The corresponding KGP logic is to be added later.

KT-79631