[K/JS] Fix source paths in source map files with per-file granularity
We had a bug with emitting of incorrect relative path to the source file
in the generated source map for per-file granularity.

The root cause of the bug was the fact that we used to calculate
the relative path from the build directory (which worked for
whole-program and per-module, since those granularities don't emit any
new directories, but for per-file we do generate new directories, so
the relative path should be calculated individually for each file).

^KT-72833 Fixed