[Scripting] Mark IrReplSnippet.*FromOtherSnippets fields as non-child
Those lists reference declarations that belong to a different repl
snippet, and are not defined (contained) in the current one, so should
not be considered as child elements. The declarations are also
implemented by lazy IR, which generally should not be visited and
lowered, but currently is because of acceptChildren and
transformChildren methods.

#KT-79807 Fixed