[AA] KT-74867 Move filtering local declarations from `KotlinAnnotationsResolver` into `LLFirIdePredicateBasedProvider`
The problem described in the KT-74867
is caused by the fact that the IDE
implementation of `KotlinAnnotationsResolver`
(see `IdeKotlinAnnotationsResolver`) does not
filter local classes by itself.

Since `LookupPredicate`s should never match
local declarations, this resulted in exceptions in the IDE
from lazy resolution because resolving local declarations
requires resolving bodies,
and that's too early for compiler plugins.

It seems like moving this logic for filtering local declarations
from the `KotlinAnnotationsResolver` implementations
into the `LLFirIdePredicateBasedProvider` is more consistent,
because this class interacts directly with `LookupPredicate`
and should obey the contract.

^KT-74867 Fixed