Reflection: do not use descriptors in KClass.typeParameters
There's a conceptual cycle when computing class type parameters: when
we're converting upper bounds, we sometimes need to access type
parameters of the same class, so they have to be already created and
accessible from somewhere. That's why `toKType` in ConvertFromJava now
takes an additional map of type parameters. Same for `toKType` in
ConvertFromMetadata, but that one already took `TypeParameterTable`.

Another possible solution here would be to do it inside the lazy value
computation in `KClass.typeParameters`: on the first access, create the
type parameter instances and store them, and then run the usual
`toKType` for all upper bounds. This way, `toKType` won't need
additional parameters. I did not choose this approach because in that
case, type parameters computation should be thread-safe, and it's not
immediately obvious which thread safety solution works best.

 #KT-75463