FIR2IR: fix exception during recursive lazy annotation constructor creation
This commit handles a situation with two very similar tests.
The test 'recursiveTypeAnnotationWithVararg.kt' reproduces KT-80885
in 2.2.20 branch, but in master KT-80940 reproduces instead and
KT-80885 is not reproducible even after this fix,
so we count KT-80885 obsolete.
KT-80940 itself is reproducible in 2.1+ only and is fixed by this commit.

The idea of the fix is to cache constructor symbol in FIR2IR before
we started to handle constructor parameters that can reference
the same constructor in certain situations (like these tests).

#KT-80940 Fixed
#KT-80885 Obsolete