[Klib] Inline IrOperation to IrExpression
To eliminate the overhead of nested Protobuf messages.

Also, reorder the expressions so that the ones that appear most
frequently occupy the first 15 indices. This is because in Protobuf, the
tags with index 0-15 require 1 byte to serialize, while subsequent
indices require 2 bytes or more.

This change reduces the size of Klibs by ~5%.

KT-84400