FE: Add expected type as EQ constraint for operators at assignments RHS
Thus preventing unexpected too strict casts to more precise types
than needed.

For constructions like `val x: A = when { .. }`, we add
`Kv === A` constraint to the synthetic call for this when.

While generally, it looks inconsistent and doesn't really make sense
(for calls we would add only `Kv <: A` constraint), it helps to
force inferring potentially less precise `A` for the whole
when-expression (see the fix for KT-65882).

Though, we specifically skipped this rule for assignments because their
RHSs might be more precise and providing information for smart casts.
But that actually led to the same problems as being solved at KT-65882
just in case of assignments (see fixed issues).

But not to lose smart cast information, we compute it as a CST
of branches and store at a dedicated type attribute used only in DFA.

This solution doesn't look less hacky than the existing solution,
but hopefully it helps with the issues and hopefully might be overcome
via KT-81083, which assumes a more general approach.

^KT-77008 Fixed
^KT-78127 Fixed
^KT-74588 Fixed
^KT-80208 Fixed