K2/LL/Stubs: never deserialize Foo.T as a type parameter
Before this commit, `class Foo<T : Foo.T>` was deserialized via stubs
to effectively `class Foo<T : T>`, which is obviously incorrect
and provoked an exception in the compiler. This commit fixes this logic,
as a type with a qualifier can never reference a type parameter.

A newly added test directly reproduces exception from KT-77082.

#KT-77082 Fixed