[WASM] Don't generate debug info on sets to temporary variables
This change is required to avoid break in debug info
in the next commit. Currently, on evaluating
 an argument of an inline function, it is wrapped
to a block without debug info.
This is not intended usage. In fact,
temporary variables should be transparent
to debug info, as they are not part of the code,
but rather an implementation detail.

After this change, workaround in inliner is no more
 needed and would be removed in the next commit.