[K/N][IR][codegen] Handled optimized away casts more gracefully
Before this commit during the autoboxing phase the optimized away casts were handled ad-hoc,
covering a number of cases. In some corner cases (probably involving interop), a box/unbox operation
might have not been inserted, leading to a crash at runtime (e.g., the UP Realm).

This commit fixes this by doing it in a completely different fashion - since IMPLICIT_CAST is a no-op
at codegen, the optimized away cast can simply be replaced with it and then handled in a general way
(while ensuring that no additional casts are introduced). This general handling ensures that all the
necessary box/unbox operations are going to be added.

 #KT-81057 Fixed