Arguments: remove tests on JVM target message
The test cannot work as written because adding a new `valueInVersions`
entry results in unstable hash code, because `KotlinReleaseVersion` is
an enum, and enums do not have stable hash code.

Also, it's questionable that `-jvm-target`/`-Xjdk-release` arguments
specifically are tested for description history, but not the rest (i.e.
all) of the arguments.

 #KT-81077