[PowerAssert] Embed immutable collections explicitly in plugin artifact
Power-Assert uses immutable collections, but it uses them as a
transitive dependency from the compiler. However, this library is shaded
in the embeddable version of the compiler. This means when running with
the non-embeddable compiler (in the IDE, for example) the power-assert
compiler plugin with throw an exception.

By explicitly including the immutable collections library as a shaded
dependency with the power-assert compiler plugin, we avoid the shading
different of the compiler. See KT-76812 for more details.

^KT-80429 Fixed