KT-74411: Implement UUID v7 generation
Implemented API to generate UUID Version 7 (as well as added a function for generating Version 4, but with a more explicit name, compared to existing random).
The implementation generates strictly monotonic uuids by utilizing a fixed-with counter, which reside in all 12bits of the rand_a field.
The counter is re-initialized every time a new timestamp is observed, an incremented every time a clock returns a non-monotonic timestamp.
In case of counter overflow, the implementation increments the timestamp and reseeds the counter.

^KT-74411 fixed

Merge-request: KT-MR-22840
Merged-by: Filipp Zhinkin <filipp.zhinkin@jetbrains.com>