[K/N][IR] Fixed a bug in casts optimization algorithm
Consider the code snippet:
    when {
        a?.f != null -> { .. }
        else -> { .. }
    }
The algorithm wrongly assumed that if a?.f is null, then a is also null.
However, if f is of a nullable type, this assumption is wrong - a might be not null,
while a.f is null). This could lead to some casts being incorrectly removed later.

 #KT-81064 Fixed