[Build] Pass `teamcity` system property to code generators
The regular way to reproduce TC environment is to pass `-Pteamcity=true`
to the gradle. But this property is a gradle property, not a system
property. So just passed this way it wouldn't be visible from the actual
runtime of task (e.g. test generator). And since our generator depend on
the condition if it's TC build or not, it makes sense to pass this
property as a system property for a convenient experience.