[FIR] Get rid of `registerDefaultComponents` in session construction
Before this commit we had the following logic:
- at first we registered default implementation of components on all
  platforms
- then on some platforms they were rewritten with platform-specific ones

But in scope of KT-78443 there will be introduced the concept of composing
multiple components, so there would be a difference between "registered
default and platform-specific one" and "registered only platform-specific
one". To express this difference all default implementations are moved
directly to component accessors.