[IR] Store preprocessed functions as private declarations
It is important to modify preprocessed inline functions to be
private. These functions are not supposed to be used anywhere.
Private declarations have a separate type of signatures.

If the signature remains public, then we can have clashes with
original functions. For example
1. Type parameter clash. If a type argument is used in the `typeOf`
call, then we handle it specially and keep it as is
(instead of erasure). This type argument must point to the type
parameter of the original function.
2. Reference to the inline function. If inside a preprocessed function
we have a reference to another inline function, then this reference
must point to the original one.

#KT-81070
#KT-81470

(cherry picked from commit 4fe4c6afea854f71b98e9b264272ca26e85c71d9)