KT-78089 Iterative matching for trivial (eagerly) quantified groups
Quantified groups matching, in general, requires
recursion, as it's the only way to correctly backtrack.
However, for some groups, named trivial here,
it is possible to implement matching iteratively,
without a recursion.
The criteria are:
- a group should not contain nested capturing groups
- a groups should not require non-trivial backtracking
- a group should not require backtracking (not contain lookarounds)

Non-trivial backtracking means that one of
the sub-expressions within a groups could be evaluated
in multiple different ways, if its successor didn't match.
That happens, for example, with alternations (a|b|c|...),
and quantifiers (nested within this group,
with the exception of a fixed number of repetitions, like {2}).

This change addresses the issue with deep recursion when matching
eagerly quantified groups.
Possessive quantifiers don't have this issue,
and reluctant quantifiers will be optimized as
a follow-up.