[FIR] Refactor builtin operations map generator into multiple functions
This is now also closer to how the backend builtin map generator is
written and easier to read.

The base datatype used is now also the Operation data class instead of
unnamed Tuple and Tripple.
The Operation doesn't use KotlinType to represent types as there aren't
any simple constructors available to represent unsigned types and this
brings it closer to the backend builtin map generator.

KT-51065

# Conflicts:
#	generators/evaluate/GenerateOperationsMap.kt