[K/N] Postpone reporting global usages with direct CCall mode to linkage
Currently, when a cinterop klib is produced with `-Xccall-mode direct`,
accessing C globals through this klib is not supported yet (KT-79742),
and therefore the global accessors are marked with `@Deprecated(ERROR)`.

This commit changes this approach: it generates a `@CCall.Direct`
annotation on a global accessor, but specifies an unresolvable symbol
name which serves as an error message to be emitted by the linker
(e.g. ld).

Motivation:
Usages of such globals are reported not during compilation but during
linkage, along with other linkage failures caused by
`-Xccall-mode direct` (e.g. usages of header-defined functions),
allowing a user to collect more problems at once.

As a nice side effect, if such a usage is unreachable, it doesn't make
the compilation fail.

^KT-81312 Fixed