[Build] Use the same compiler version in all Gradle-related modules
Previously, it was set up to use outdated compiler for main-like compilations to be able to use outdated AV/LV for Gradle compatibility.
To simplify the maintenance, get rid of customized compiler versions within a project and use the same across all source sets and compilations. With this change, main-like compilation are still configured to use outdated AV/LV, but the rest is using the default AV/LV for that outdated compiler.
Relates to KT-78200