[K/N][build] Use relative path for install_name in NativePlugin
By default, install_name in a dylib is set from the output path.
If the output is specified as the absolute path, the absolute path
will be recorded, which will make the output non-reproducible.
Setting output to a relative path fixes the issue. However,
this is very implicit and feels brittle. So, in this commit
install_name is set explicitly via a linker flag.

^KT-81500