[Gradle] Dont add parcelize plugin to JVM compilations
* Added better heuristic for External Android Target
Using android name when external target ecnountered should be safe enough
however Google may change it in the future so this is still heuristic
And propper integration with parcelize plugin is required. see KT-81384

* Added functional test to verify that parcelize is not added to jvm
targets and is added to legacy android target
Adding simillar functional test for External Target is not possible yet
because KGP is compiled against old AGP (8.2) and parcelize is not
supported in this version.
Instead Integration test is added for AGP 8.11.

^KT-81384
^KT-81249 Verification Pending