Reflection: remove obsolete code about local delegated properties
Before Kotlin 1.5 where new JVM IR backend was enabled by default, local
delegated properties declared in interfaces had DefaultImpls class as a
container. This code was needed because metadata of local delegated
properties is stored in the interface (DefaultImpls is a synthetic class
and has no meaningful metadata).

Note that this means it won't be possible anymore to load local
delegated properties in interfaces compiled with Kotlin <1.5 with
kotlin-reflect 2.3.20+, which is probably not a big deal.

Also, copy the test with JVM default mode "disable" to keep testing the
original issue.

 #KT-80710