JVM: add language feature for loading annotations on property annotations
This logic is needed to enable `LanguageFeature.AnnotationsInMetadata`
by default (KT-75736) without causing a behavior change that is
described in a comment in KT-22463 and is checked in the newly added
tests.

The feature is disabled by default, and it will be enabled only when
KT-22463 is fixed (with the Language Committee approval).

 #KT-22463